+++ contrib/opencensus-ext-azure/opencensus/ext/azure/common/__init__.py
@@ -46,7 +46,14 @@ def process_options(options):
     endpoint = code_cs.get(INGESTION_ENDPOINT) \
         or env_cs.get(INGESTION_ENDPOINT) \
         or 'https://dc.services.visualstudio.com'
-    options.endpoint = endpoint + '/v2/track'
+    options.endpoint = endpoint
+
+    # Authorization
+    # `azure.core.credentials.TokenCredential` class must be valid
+    if options.credential and not hasattr(options.credential, 'get_token'):
+        raise ValueError(
+            'Must pass in valid TokenCredential.'
+        )
 
     # storage path
     if options.storage_path is None:
@@ -101,13 +108,15 @@ class Options(BaseObject):
 
     _default = BaseObject(
         connection_string=None,
+        credential=None,  # Credential class used by AAD auth
         enable_local_storage=True,
-        enable_standard_metrics=True,
+        enable_standard_metrics=True,  # Used by metrics exporter, True to send standard metrics  # noqa: E501
+        enable_stats_metrics=True,  # True to send stats metrics
         endpoint='https://dc.services.visualstudio.com/v2/track',
         export_interval=15.0,
         grace_period=5.0,
         instrumentation_key=None,
-        logging_sampling_rate=1.0,
+        logging_sampling_rate=1.0,  # Used by log exporter, controls sampling
         max_batch_size=100,
         minimum_retry_interval=60,  # minimum retry interval in seconds
         proxies=None,  # string maps url schemes to the url of the proxies