+++ contrib/opencensus-ext-azure/opencensus/ext/azure/common/transport.py
@@ -14,13 +14,23 @@
 
 import json
 import logging
+import threading
 
 import requests
+from azure.core.exceptions import ClientAuthenticationError
+from azure.identity._exceptions import CredentialUnavailableError
 
 logger = logging.getLogger(__name__)
+_MONITOR_OAUTH_SCOPE = "https://monitor.azure.com//.default"
+_requests_lock = threading.Lock()
+_requests_map = {}
 
 
 class TransportMixin(object):
+    def _check_stats_collection(self):
+        return self.options.enable_stats_metrics and \
+            (not hasattr(self, '_is_stats') or not self._is_stats)
+
     def _transmit_from_storage(self):
         if self.storage:
             for blob in self.storage.gets():
@@ -45,13 +55,22 @@ class TransportMixin(object):
         if not envelopes:
             return 0
         try:
+            headers = {
+                'Accept': 'application/json',
+                'Content-Type': 'application/json; charset=utf-8',
+            }
+            endpoint = self.options.endpoint
+            if self.options.credential:
+                token = self.options.credential.get_token(_MONITOR_OAUTH_SCOPE)
+                headers["Authorization"] = "Bearer {}".format(token.token)
+                # Use new api for aad scenario
+                endpoint += '/v2.1/track'
+            else:
+                endpoint += '/v2/track'
             response = requests.post(
-                url=self.options.endpoint,
+                url=endpoint,
                 data=json.dumps(envelopes),
-                headers={
-                    'Accept': 'application/json',
-                    'Content-Type': 'application/json; charset=utf-8',
-                },
+                headers=headers,
                 timeout=self.options.timeout,
                 proxies=json.loads(self.options.proxies),
             )
@@ -59,11 +78,22 @@ class TransportMixin(object):
             logger.warning(
                 'Request time out. Ingestion may be backed up. Retrying.')
             return self.options.minimum_retry_interval
-        except Exception as ex:  # TODO: consider RequestException
+        except requests.RequestException as ex:
             logger.warning(
                 'Retrying due to transient client side error %s.', ex)
             # client side error (retryable)
             return self.options.minimum_retry_interval
+        except CredentialUnavailableError as ex:
+            logger.warning('Credential error. %s. Dropping telemetry.', ex)
+            return -1
+        except ClientAuthenticationError as ex:
+            logger.warning('Authentication error %s', ex)
+            return self.options.minimum_retry_interval
+        except Exception as ex:
+            logger.warning(
+                'Error when sending request %s. Dropping telemetry.', ex)
+            # Extraneous error (non-retryable)
+            return -1
 
         text = 'N/A'
         data = None
@@ -77,6 +107,9 @@ class TransportMixin(object):
             except Exception:
                 pass
         if response.status_code == 200:
+            if self._check_stats_collection():
+                with _requests_lock:
+                    _requests_map['success'] = _requests_map.get('success', 0) + 1  # noqa: E501
             return 0
         if response.status_code == 206:  # Partial Content
             if data:
@@ -119,6 +152,24 @@ class TransportMixin(object):
                 text,
             )
             # server side error (retryable)
+            return self.options.minimum_retry_interval
+        # Authentication error
+        if response.status_code == 401:
+            logger.warning(
+                'Authentication error %s: %s.',
+                response.status_code,
+                text,
+            )
+            return self.options.minimum_retry_interval
+        # Forbidden error
+        # Can occur when v2 endpoint is used while AI resource is configured
+        # with disableLocalAuth
+        if response.status_code == 403:
+            logger.warning(
+                'Forbidden error %s: %s.',
+                response.status_code,
+                text,
+            )
             return self.options.minimum_retry_interval
         logger.error(
             'Non-retryable server side error %s: %s.',