+++ contrib/opencensus-ext-azure/opencensus/ext/azure/log_exporter/__init__.py
@@ -30,6 +30,7 @@ from opencensus.ext.azure.common.protocol import (
 )
 from opencensus.ext.azure.common.storage import LocalFileStorage
 from opencensus.ext.azure.common.transport import TransportMixin
+from opencensus.ext.azure.metrics_exporter import statsbeat_metrics
 from opencensus.trace import execution_context
 
 logger = logging.getLogger(__name__)
@@ -61,6 +62,11 @@ class BaseLogHandler(logging.Handler):
         self._queue = Queue(capacity=self.options.queue_capacity)
         self._worker = Worker(self._queue, self)
         self._worker.start()
+        # start statsbeat on exporter instantiation
+        if self.options.enable_stats_metrics:
+            statsbeat_metrics.collect_statsbeat_metrics(
+                self.options.instrumentation_key
+            )
 
     def _export(self, batch, event=None):  # pragma: NO COVER
         try: